config: add modularization, gitignore and settingsjson
