+++ .vscode/settings.json
@@ -1,0 +1,9 @@
+{
+    "files.exclude": {
+        "**/*.pyc": {
+            "when": "$(basename).py"
+        },
+        "**/__pycache__": true,
+        "**/*.pytest_cache": true,
+    }
+}
