+++ src/models/job.py
@@ -1,0 +1,10 @@
+# class Job(BaseModel):
+#     title:str
+#     company:str
+#     location:str
+#     time_opened:str
+#     company_img:str
+#     link:str
+        
+    
+
